requirements.txt
pytesseract>=0.3.10  # OCR (Optical Character Recognition) Engine für Texterkennung in Bildern
opencv-python>=4.8.0  # Computer Vision Bibliothek für Bildverarbeitung und -analyse
numpy>=1.24.0  # Numerische Berechnungen und Array-Operationen
Pillow>=10.0.0  # Python Imaging Library für Bildverarbeitung
mlflow>=2.0.0  # Machine Learning Experiment Tracking und Model Management
pdf2image>=1.16.0  # Konvertierung von PDF-Dateien in Bilder
